Refactor file upload and analytics
